# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lotus Driver Tracking System",
    "description": "Backend for driver signup/login, barcode-scan order registration and delivery completion (GPS + timestamp logged per scan), and an SSO-gated admin dashboard for reviewing driver activity.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/scans/register": {
      "post": {
        "summary": "Register an order by its scanned barcode into the given job (started via POST /api/manifests/start); a repeat scan of the same code in that job is a no-op, not an error.",
        "operationId": "registerScan",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScanRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Registered (or already was).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScanRegisterResult" } } }
          },
          "404": { "description": "Job not found -- start one first with POST /api/manifests/start." },
          "409": { "description": "That code was already delivered or was cancelled, or the job was cancelled." },
          "422": { "description": "Scanned code is empty, or manifest_id is missing." }
        }
      }
    },
    "/api/scans/complete": {
      "post": {
        "summary": "Mark an order delivered by scanning the same barcode again, with a required proof photo; logs a delivered event with timestamp + GPS + photo. If the order was never scanned in at the warehouse, it's auto-registered into today's most recent open job instead of being blocked (no 'registered' event/timestamp for it). Response flags job_complete if this resolved the last open order in its job.",
        "operationId": "completeScan",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": { "type": "string", "description": "The decoded barcode value." },
                  "lat": { "type": "number" },
                  "lng": { "type": "number" },
                  "occurred_at": { "type": "string", "format": "date-time", "description": "ISO 8601; defaults to server time if omitted." },
                  "photo": { "type": "string", "format": "binary", "description": "Required proof-of-delivery photo." }
                },
                "required": ["code", "photo"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marked delivered.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScanCompleteResult" } } }
          },
          "404": { "description": "No job is open today to auto-register this code into -- tap \"Arrived at warehouse\" first." },
          "409": { "description": "That code was already delivered or was cancelled." },
          "422": { "description": "Scanned code is empty." }
        }
      }
    },
    "/api/scans/fail": {
      "post": {
        "summary": "Mark a delivery attempt failed by scanning the barcode, with a required reason and proof photo; logs a failed event with timestamp + GPS + reason + photo. If the order was never scanned in at the warehouse, it's auto-registered into today's most recent open job instead of being blocked (no 'registered' event/timestamp for it). Response flags job_complete if this resolved the last open order in its job.",
        "operationId": "failScan",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": { "type": "string", "description": "The decoded barcode value." },
                  "reason": { "type": "string", "description": "Why the delivery attempt failed." },
                  "lat": { "type": "number" },
                  "lng": { "type": "number" },
                  "occurred_at": { "type": "string", "format": "date-time", "description": "ISO 8601; defaults to server time if omitted." },
                  "photo": { "type": "string", "format": "binary", "description": "Required proof photo of the failed attempt." }
                },
                "required": ["code", "reason", "photo"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marked failed.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScanCompleteResult" } } }
          },
          "404": { "description": "No job is open today to auto-register this code into -- tap \"Arrived at warehouse\" first." },
          "409": { "description": "That code was already delivered or was cancelled." },
          "422": { "description": "Scanned code is empty, or reason is empty." }
        }
      }
    },
    "/api/driver/warehouse": {
      "put": {
        "summary": "Set which outlet this driver is assigned to -- a fixed assignment (not per-job), changeable any time from the driver's profile screen.",
        "operationId": "updateDriverWarehouse",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DriverWarehouseRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Updated driver info.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "email": { "type": "string" },
                        "name": { "type": "string" },
                        "warehouse_id": { "type": "integer", "nullable": true },
                        "warehouse_name": { "type": "string", "nullable": true }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": { "description": "That outlet doesn't exist or is no longer active." }
        }
      }
    },
    "/api/manifests/start": {
      "post": {
        "summary": "\"Arrived at warehouse\" -- always starts a brand-new job (a driver may make more than one warehouse trip a day), timestamped + geotagged at creation. Every order scanned afterwards groups into this job until another one is started.",
        "operationId": "startManifest",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArrivalRequest" } } }
        },
        "responses": {
          "201": {
            "description": "The newly created job.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "manifest": { "$ref": "#/components/schemas/Manifest" } } }
              }
            }
          }
        }
      }
    },
    "/api/manifests": {
      "get": {
        "summary": "List the current driver's scan sessions, most recent work date first.",
        "operationId": "listManifests",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "responses": {
          "200": {
            "description": "Session list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "manifests": { "type": "array", "items": { "$ref": "#/components/schemas/Manifest" } } }
                }
              }
            }
          }
        }
      }
    },
    "/api/manifests/{manifest_id}": {
      "get": {
        "summary": "Get one scan session belonging to the current driver, with its registered/delivered jobs.",
        "operationId": "getManifest",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "parameters": [{ "name": "manifest_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Session and its jobs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "manifest": { "$ref": "#/components/schemas/Manifest" },
                    "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/Job" } }
                  }
                }
              }
            }
          },
          "404": { "description": "Session not found." }
        }
      }
    },
    "/api/manifests/{manifest_id}/cancel": {
      "post": {
        "summary": "Cancel a scan session (e.g. the wrong load got scanned in) so the driver can start clean; refused once any of its jobs has been delivered.",
        "operationId": "cancelManifest",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "parameters": [{ "name": "manifest_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Session cancelled; its still-registered jobs are marked cancelled too.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "manifest": { "$ref": "#/components/schemas/Manifest" } } }
              }
            }
          },
          "404": { "description": "Session not found." },
          "409": { "description": "Already cancelled, or has completed deliveries and can't be cancelled." }
        }
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "summary": "List the chronological scan events (registered, delivered) for one of the current driver's jobs.",
        "operationId": "listJobEvents",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "parameters": [{ "name": "job_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Event list, oldest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } }
                }
              }
            }
          },
          "404": { "description": "Job not found." }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "driverBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Driver JWT from /api/auth/signup or /api/auth/login, signed with the platform's JWT_SECRET."
      },
      "adminSSO": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Forwarded-Email",
        "description": "Injected by the Substrait platform's Google SSO proxy when SSO is enabled for this app in the portal's Access tab; not client-supplied. The caller must also be on the admin allowlist in the users table."
      }
    },
    "schemas": {
...
      "ScanRequest": {
        "type": "object",
        "required": ["code", "manifest_id"],
        "properties": {
          "code": { "type": "string", "description": "The decoded barcode value." },
          "manifest_id": { "type": "integer", "description": "The job to register this order into (from POST /api/manifests/start)." },
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "occurred_at": { "type": "string", "format": "date-time", "nullable": true, "description": "ISO 8601; defaults to server time if omitted." }
        }
      },
      "ArrivalRequest": {
        "type": "object",
        "properties": {
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "occurred_at": { "type": "string", "format": "date-time", "nullable": true, "description": "ISO 8601; defaults to server time if omitted." }
        }
      },
...
      "DriverWarehouseRequest": {
        "type": "object",
        "required": ["warehouse_id"],
        "properties": { "warehouse_id": { "type": "integer" } }
      },
...
      "ScanRegisterResult": {
        "type": "object",
        "properties": {
          "job_id": { "type": "integer" },
          "manifest_id": { "type": "integer" },
          "tracking_no": { "type": "string" },
          "already_registered": { "type": "boolean" }
        }
      },
      "ScanCompleteResult": {
        "type": "object",
        "properties": {
          "job_id": { "type": "integer" },
          "tracking_no": { "type": "string" },
          "manifest_id": { "type": "integer" },
          "job_complete": { "type": "boolean", "description": "True if every order in this job now has a resolved outcome." }
        }
      },
      "Manifest": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "work_date": { "type": "string", "format": "date" },
          "cancelled_at": { "type": "string", "nullable": true },
          "warehouse_arrived_at": { "type": "string", "nullable": true },
          "warehouse_arrived_lat": { "type": "number", "nullable": true },
          "warehouse_arrived_lng": { "type": "number", "nullable": true },
          "created_at": { "type": "string" }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tracking_no": { "type": "string" },
          "status_code": { "type": "string", "enum": ["registered", "delivered", "failed", "cancelled"] },
          "created_at": { "type": "string" }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "status_code": { "type": "string", "enum": ["registered", "delivered", "failed"] },
          "occurred_at": { "type": "string" },
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "failure_reason": { "type": "string", "nullable": true },
          "photo_id": { "type": "integer", "nullable": true }
        }
      },
...
    }
  }
}